requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
ta>=0.7.0  # Pour les indicateurs techniques
kaggle==1.5.16
requests==2.31.0
python-dotenv==1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2 